requirements.txt
torch>=2.0
numpy>=1.23
opencv-python>=4.8
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
trimesh>=4.0
